관리자 수정/삭제 불가능한 이슈 해결

+++ src/main/java/com/sparta/myBlogBackendServer/service/CommentService.java
@@ -4,6 +4,7 @@ import com.sparta.myBlogBackendServer.dto.CommentRequestDto;
 import com.sparta.myBlogBackendServer.entity.Comment;
 import com.sparta.myBlogBackendServer.entity.Post;
 import com.sparta.myBlogBackendServer.entity.User;
+import com.sparta.myBlogBackendServer.entity.UserRoleEnum;
 import com.sparta.myBlogBackendServer.repository.CommentRepository;
 import com.sparta.myBlogBackendServer.repository.PostRepository;
 import com.sparta.myBlogBackendServer.repository.UserRepository;
@@ -29,26 +30,45 @@ public class CommentService {
         User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
                 () -> new IllegalArgumentException("유저 정보가 존재하지 않습니다.")
         );
+
         Post post = postRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("게시글이 존재하지 않습니다.")
         );
+
         Comment comment = new Comment(commentRequestDto, post, user);
         commentRepository.save(comment);
     }
 
     @Transactional
-    public void updateComment(Long commentId, CommentRequestDto commentRequestDto) {
+    public void updateComment(Long commentId, CommentRequestDto commentRequestDto, Claims claims) {
+        User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                () -> new IllegalArgumentException("유저 정보가 존재하지 않습니다.")
+        );
         Comment comment = commentRepository.findById(commentId).orElseThrow(
                 () -> new IllegalArgumentException("댓글이 존재하지 않습니다.")
         );
-        comment.update(commentRequestDto);
-        commentRepository.save(comment);
+
+        UserRoleEnum role = user.getRole();
+        if (user.getUsername().equals(comment.getUser().getUsername()) || role.equals(UserRoleEnum.ADMIN)) {
+            comment.update(commentRequestDto);
+            commentRepository.save(comment);
+        } else {
+            throw new IllegalArgumentException("접근할 수 있는 권한이 없습니다.");
+        }
     }
 
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, Claims claims) {
+        User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                () -> new IllegalArgumentException("유저 정보가 존재하지 않습니다.")
+        );
         Comment comment = commentRepository.findById(commentId).orElseThrow(
                 () -> new IllegalArgumentException("댓글이 존재하지 않습니다.")
         );
-        commentRepository.delete(comment);
+        UserRoleEnum role = user.getRole();
+        if (user.getUsername().equals(comment.getUser().getUsername()) || role.equals(UserRoleEnum.ADMIN)) {
+            commentRepository.delete(comment);
+        } else {
+            throw new IllegalArgumentException("접근할 수 있는 권한이 없습니다.");
+        }
     }
 }
